ethtool: fec: add pretty dump
Add pretty dump for the port registers of the "fec" kernel driver. Only
offsets exposed by the driver are dumped by ethtool.

Both register dump versions are supported, version 2 which corresponds
to the kernel compilation test:

    #if defined(CONFIG_M523x) || defined(CONFIG_M527x)
    || defined(CONFIG_M528x) || defined(CONFIG_M520x)
    || defined(CONFIG_M532x) || defined(CONFIG_ARM)
    || defined(CONFIG_ARM64) || defined(CONFIG_COMPILE_TEST)

and version 1 which corresponds to the opposite test. At the same time,
detail a few interesting registers of version 2.

Kernels not patched for setting this version will cause ethtool to
dump the whole set of registers as it already does today.

Signed-off-by: Vivien Didelot <vivien.didelot@gmail.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>