ethtool: add 10000baseR_FEC link mode name
ETHTOOL_LINK_MODE_10000baseR_FEC_BIT link mode is listed in
ethtool_link_mode_bit_indices array but doesn't have its name assigned in
the table in dump_link_caps() (so that it's not shown in "ethtool <dev>"
output) and is not listed in the manual page. Add it to both.

Signed-off-by: Michal Kubecek <mkubecek@suse.cz>
Signed-off-by: John W. Linville <linville@tuxdriver.com>