netlink: Add --cable-test action
This action is used to trigger an Ethernet cable test.  Since this is
a long operation, taking 7 seconds or more, the results are not
synchronously reported. Rather the results are reported as an event
sometime later. The same code used by monitor to print events is then
used to print the results of the cable test.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>