ethtool: igb: dump RR2DCDELAY register
Decode 'RR2DCDELAY' register which Linux kernel provides starting
from version 5.3. The corresponding commit in the Linux kernel is:
    cd502a7f7c9c igb: add RR2DCDELAY to ethtool registers dump

The RR2DCDELAY register is present in I210 and I211 Intel Gigabit
Ethernet chips and it stands for "Read Request To Data Completion
Delay". Here is how this register is described in the I210 datasheet:

"This field captures the maximum PCIe split time in 16 ns units,
which is the maximum delay between the read request to the first data
completion. This is giving an estimation of the PCIe round trip time."

In practice, this register can be used to measure the time it takes
the NIC to read data from the host memory.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>
Reviewed-by: Michal Kubecek <mkubecek@suse.cz>
Signed-off-by: John W. Linville <linville@tuxdriver.com>