netlink: add handler for permaddr (-P)
Implement "ethtool -P <dev>" subcommand. This retrieves and displays
information traditionally provided by ETHTOOL_GPERMADDR ioctl request.

Permanent hardware address is provided by rtnetlink RTM_GETLINK request
so that basic support for rtnetlink (which is not based on genetlink)
requests is also added.

The IFLA_PERM_ADDRESS attribute in RTM_NEWLINK message is not set if the
device driver did not set its permanent hardware address which is hard to
distinguish from ethtool running on top of an older kernel not supporting
the attribute. As IFLA_PERM_ADDRESS attribute was added to kernel shortly
before the initial portion of ethtool netlink interface, we let ethtool
initialize ethtool netlink even if it is not needed so that we can fall
back to ioctl on older kernels and assume that absence of IFLA_PERM_ADDRESS
means the permanent address is not set. This would be only wrong on some
distribution kernel which would backport ethtool netlink interface but not
IFLA_PERM_ADDRESS which is unlikely.

Permanent address is immutable (or not mutable using standard interfaces)
so that there is no related notification message.

Signed-off-by: Michal Kubecek <mkubecek@suse.cz>
Signed-off-by: John W. Linville <linville@tuxdriver.com>