netlink: message format descriptions for rtnetlink
Add description of rtnetlink message formats to be used for pretty printing
infrastructure and helper function to pretty print rtnetlink messages.

As rtnetlink is quite complex and only RTM_GETLINK and RTM_NEWLINK messages
are interesting for ethtool at the moment, this commit provides only names
of a limited subset of rtnetlink messages and names and formats of top
level attributes used in RTM_{NEW,DEL,GET,SET}LINK messages.

Signed-off-by: Michal Kubecek <mkubecek@suse.cz>
Signed-off-by: John W. Linville <linville@tuxdriver.com>