ethtool: use "Not reported" when no FEC modes are provided
When displaying the FEC link modes advertised by the peer, we used the
string "No" to indicate when nothing was provided. This does not match
the IOCTL output which indicates "Not reported". It also doesn't match
the local advertised FEC modes, which also used the "Not reported"
string.

This is especially confusing for FEC, because the FEC bits include
a "None" bit which indicates that FEC is definitely not supported. Avoid
this confusion and match both the local advertised settings display and
the old IOCTL output by using "Not reported" when FEC settings aren't
reported.

Fixes: 10cc3ea337d1 ("netlink: partial netlink handler for gset (no option)")
Reported-by: Andrew Lunn <andrew@lunn.ch>
Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Michal Kubecek <mkubecek@suse.cz>