netlink: Fix the condition for displaying actual changes
This comment in the code:

    /* result is not exactly as requested, show differences */

implies that the "Actual changes" output should be displayed only if the
result is not as requested, which matches the legacy ethtool behavior.
However, in fact, ethtool-netlink displays "actual changes" even when
the changes are expected (e.g., one bit was requested, and it was
changed as requested).

This commit fixes the condition above to make the behavior match the
description in the comment and the behavior of the legacy ethtool. The
new condition excludes the req_mask bits from active_mask to avoid
reacting on bit changes that we asked for. The new condition now
matches the ifs in the loop above that print "[requested on/off]" and
"[not requested]".

Signed-off-by: Maxim Mikityanskiy <maximmi@mellanox.com>
Signed-off-by: Michal Kubecek <mkubecek@suse.cz>