ioctl: only memset non-NULL link settings
In commit bef780467fa ("ioctl: do not pass transceiver value back to
kernel") a regression slipped in. If we have a kernel that does not
support the ETHTOOL_xLINKSETTINGS API, then the do_ioctl_glinksettings()
function will return a NULL pointer.

Hence before memset'ing the pointer to zero we must first check it is
valid, as NULL return is perfectly fine when running on old kernels.

Fixes: bef780467fa7 ("ioctl: do not pass transceiver value back to kernel")
Signed-off-by: Hans-Christian Noren Egtvedt <hegtvedt@cisco.com>
Signed-off-by: Michal Kubecek <mkubecek@suse.cz>