sff-common: Move cache of SFP pages into common place
The decoding of SFP data needs to be common for both IOCTL and netlink
mechanics for accessing the data. By moving the cache of pages into
common code, the decoders can use that for accessing the data,
independent of how it was retrieved. Add the sff_ prefix to the
functions to keep the namespace clean.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>