Populate cache with IOCTL SFP data
Break the SFP data up from the IOCTL call into pages, and place it
into the cache. The page layout differs depending on the
modinfo.type. And the driver is not required to provide all the pages
for a type. So take both the type and the length into account when
populating the cache.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>