sff8472_show_all: Get EEPROM data from the cache
In a step towards unifying IOCTL and netlink code paths, when printing
sff8472 ID and diagnostics data, get the pages out of the cache,
rather than pass it in a buffer. sff8472 extends 8079 by adding extra
bytes in the first page to indicate compliance with 8472 and if the
diagnostics data is available. As part of the first page, this should
always be available. However the diagnostics page might not of been
provided by the kernel. So split the parsing up, and only parse what
is available.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>