sff8079_show_all: Get EEPROM data from the cache
In a step towards unifying IOCTL and netlink code paths, when printing
sff8079 ID data, get the page out of the cache, rather than pass it in
a buffer. This requires a bit of plumbing, since the context
information can be required if there is a cache miss when using
netlink.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>